Moved remaps to their own Handler class
