+++ src/main/resources/data/minecraft/tags/blocks/crops.json
@@ -1,0 +1,10 @@
+{
+  "replace": false,
+  "values": [
+    "farmersdelight:cabbages",
+    "farmersdelight:tomatoes",
+    "farmersdelight:onions",
+    "farmersdelight:rice_crop",
+    "farmersdelight:tall_rice_crop"
+  ]
+}
